+++ README.md
@@ -1,2 +1,4 @@
 # Thuto
 It is essentially some code I gathered from old project and I think I can reuse, instead of copy past it in my new projects (I did this too much sadly) I decided to make a library out of it.
+
+It required C++20.
